refactor Paths for utils
